+++ catalog/app/containers/Admin/Buckets/gql/BucketsUpdate.generated.ts
@@ -49,6 +49,7 @@ export interface containers_Admin_Buckets_gql_BucketsUpdateMutation {
           readonly fileExtensionsToIndex: ReadonlyArray<string> | null
           readonly indexContentBytes: number | null
           readonly scannerParallelShardsDepth: number | null
+          readonly prefixes: ReadonlyArray<string>
           readonly snsNotificationArn: string | null
           readonly skipMetaDataIndexing: boolean | null
           readonly lastIndexed: Date | null
@@ -170,6 +171,7 @@ export const containers_Admin_Buckets_gql_BucketsUpdateDocument = {
           { kind: 'Field', name: { kind: 'Name', value: 'fileExtensionsToIndex' } },
           { kind: 'Field', name: { kind: 'Name', value: 'indexContentBytes' } },
           { kind: 'Field', name: { kind: 'Name', value: 'scannerParallelShardsDepth' } },
+          { kind: 'Field', name: { kind: 'Name', value: 'prefixes' } },
           { kind: 'Field', name: { kind: 'Name', value: 'snsNotificationArn' } },
           { kind: 'Field', name: { kind: 'Name', value: 'skipMetaDataIndexing' } },
           { kind: 'Field', name: { kind: 'Name', value: 'lastIndexed' } },